Avoid including `include/*` when `pulp_cluster` is a submodule.
When instantiating `pulp_cluster` standalone, you want to fetch the
files inside the `include` directory, and that is handled by the
Makefile. Bender explicitly uses the `cluster_standalone` cluster to
include the `pulp_interfaces`. When `pulp_cluster` is a submodule, the
only defines you want to include, are the ones from the top module!